Move paypal and cart modal to modals folder
