+++ client/src/components/CartItems.js
@@ -1,6 +1,6 @@
 import React, {Fragment, useContext, useEffect, useState} from 'react'
 import CartItem from './CartItem'
-import Modal from "../components/Modal"
+import Modal from "./modals/Modal"
 import ShippingForm from "./ShippingForm.js"
 import AppContext from "../context/AppContext";
 import AlertContext from "../context/alertContext";
